Rename media stream to sender and reciever
