+++ ctsTraffic/ctsTraffic.vcxproj.filters
@@ -54,13 +54,13 @@
     <ClCompile Include="ctsTraffic.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ctsMediaStreamServerConnectedSocket.cpp">
+    <ClCompile Include="ctsMediaStreamSender.cpp">
       <Filter>MediaStreaming</Filter>
     </ClCompile>
     <ClCompile Include="ctsMediaStreamServerListeningSocket.cpp">
       <Filter>MediaStreaming</Filter>
     </ClCompile>
-    <ClCompile Include="ctsMediaStreamClient.cpp">
+    <ClCompile Include="ctsMediaStreamReceiver.cpp">
       <Filter>MediaStreaming</Filter>
     </ClCompile>
     <ClCompile Include="ctsWinsockLayer.cpp">
@@ -175,7 +175,7 @@
     <ClInclude Include="ctsMediaStreamProtocol.hpp">
       <Filter>MediaStreaming</Filter>
     </ClInclude>
-    <ClInclude Include="ctsMediaStreamServerConnectedSocket.h">
+    <ClInclude Include="ctsMediaStreamSender.h">
       <Filter>MediaStreaming</Filter>
     </ClInclude>
     <ClInclude Include="ctsMediaStreamServerListeningSocket.h">
@@ -184,7 +184,7 @@
     <ClInclude Include="ctsMediaStreamServer.h">
       <Filter>MediaStreaming</Filter>
     </ClInclude>
-    <ClInclude Include="ctsMediaStreamClient.h">
+    <ClInclude Include="ctsMediaStreamReceiver.h">
       <Filter>MediaStreaming</Filter>
     </ClInclude>
     <ClInclude Include="ctsWinsockLayer.h">
